requirements.txt
pydantic>=2.0.0
aiohttp>=3.8.0
numpy>=1.21.0
python-dotenv>=1.0.0
tenacity>=8.0.0
rich>=13.0.0
strands-agents
